Refactor CLI helpers into modules
