+++ crates/cli/src/commands/mod.rs
@@ -1,0 +1,8 @@
+pub mod build;
+mod events;
+pub mod init;
+pub mod report;
+pub mod run;
+pub mod status;
+
+pub(crate) use events::read_recent_events;
